# docs/swagger.yaml
basePath: /api/v1
definitions:
  admin.AdminLoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  admin.AdminLoginResponse:
    properties:
      admin_session:
        example: true
        type: boolean
      role:
        example: super_admin
        type: string
      tokens:
        $ref: '#/definitions/admin.TokensResponse'
    type: object
...
  admin.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  admin.TokensResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
...
host: qr-parking-back.abdullokh.com
info:
  contact: {}
  description: Scan a QR code to view vehicle owner info and send them a message.
  title: QR-Parking API
  version: "1.0"
paths:
...
  /admin/login:
    post:
      consumes:
      - application/json
      description: Authenticates admin with username/password and returns JWT with
        admin_session=true
      parameters:
      - description: Admin credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Admin login
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"